refactor: 최대글자수를 체크하는 validator 함수 생성
❤️
