+++ frontend/src/constants/index.js
@@ -1,7 +1,13 @@
 export { default as PATH } from './path';
 export { default as API } from './api';
 export { default as PROFILE_PAGE_MENU } from './profilePageMenu';
-export { CONFIRM_MESSAGE, ALERT_MESSAGE, PLACEHOLDER, SNACKBAR_MESSAGE } from './message';
-export { POST_TITLE } from './input';
+export {
+  CONFIRM_MESSAGE,
+  ALERT_MESSAGE,
+  PLACEHOLDER,
+  SNACKBAR_MESSAGE,
+  ERROR_MESSAGE,
+} from './message';
+export { POST_TITLE, PROFILE, CHART, REPORT_DESCRIPTION } from './input';
 export { default as COLOR } from './color';
 export { REQUEST_REPORT_TYPE } from './requestType';
