+++ frontend/src/utils/validator.ts
@@ -1,0 +1,5 @@
+export const limitLetterLength = (words: string, limitedLength: number) => {
+  if (words.length > limitedLength) return false;
+
+  return true;
+};
